lib/core/system: time_monotonic(): return f64.¹
1: To preserve FP precision over time.

+++ lib/core/src/togo/core/system/system.hpp
@@ -39,7 +39,7 @@ void sleep_ms(unsigned duration_ms);
 ///
 /// The return value should have a precision of milliseconds or
 /// better (typically nanoseconds).
-float time_monotonic();
+f64 time_monotonic();
 
 /// Get environment variable value.
 ///
